chore: move handler tests to separated directory
